feat: handle color names for deprecated
